+++ paddle/ap/include/axpr/serializable_list_method_class.h
@@ -46,7 +46,7 @@ struct MethodClassImpl<ValueT, adt::List<SerializableValue>> {
           if (index < 0) {
             index += self->size();
           }
-          if (index >= 0 && index < self->size()) {
+          if (index >= 0 && index < static_cast<int64_t>(self->size())) {
             return self->at(index).template CastTo<ValueT>();
           }
           return adt::errors::IndexError{"list index out of range"};
